PaymentMethodInfo as interface #1053
